Feat: specification details optimization
