+++ src/components/SwitchButton/index.jsx
@@ -3,7 +3,7 @@ import './index.less';
 import { Button } from '@alifd/next';
 import { useTranslation } from 'react-i18next';
 import { publish } from 'tiny-pubsub';
-import { getLanguage } from './utils';
+import { getLanguage } from '../../utils/common';
 
 const SwitchLanguage = () => {
   let { i18n } = useTranslation();
