+++ src/i18n.tsx
@@ -2,7 +2,8 @@ import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
 import En from './locals/En/en.json';
 import Zh from './locals/Zh/zh.json';
-import { getLanguage } from './components/SwitchButton/utils';
+import { getLanguage } from './utils/common';
+
 const resources = {
   en: {
     translation: En,
