+++ src/store/model/application.js
@@ -1,5 +1,5 @@
 import { getApplicationList } from '../../api/application';
-import { getAppCardList } from '../../utils/common';
+
 
 export default {
   namespace: 'application',
@@ -26,8 +26,29 @@ export default {
   effects: {
     *getApplicationList(action, { call, put }) {
       const result = yield call(getApplicationList);
-      const appContent = getAppCardList(result.data);
+      const appContent = getAppCardList(result.data || {});
       yield put({ type: 'updateApplicationList', payload: appContent });
     },
   },
 };
+
+function getAppCardList(data) {
+  const applicationsList = data.applications;
+  const appContent = [];
+  for (const item of applicationsList) {
+    const rules = item.gatewayRule && item.gatewayRule[0];
+    const { protocol, address, componentPort } = rules;
+    const href = protocol + address + componentPort;
+    const app = {
+      name: item.name,
+      status: item.status,
+      icon: item.icon,
+      description: item.description,
+      createTime: item.createTime,
+      btnContent: item.btnContent,
+      href: href,
+    };
+    appContent.push(app);
+  }
+  return appContent;
+}
